apps/backend/requirements.txt
fastapi
pydantic
supabase
PyJWT
requests
uvicorn[standard]
python-dotenv